chore: improved add form dialog
